fix: Update unit test checkpoint files to match the new InMemoryStorage structure in Optuna 4.3.0.
* Fix: Update unit test checkpoint files to match the new InMemoryStorage structure in Optuna 4.3.0.

* update changelog and pyproject

* pyproject

---------

+++ tests/tuning/test_end_to_end.py
@@ -131,5 +131,7 @@ def test_tune(tmpdir, n_trials, two_phase_tuning, start_from_checkpoint):
         else:
             assert_results(output_dir)
     finally:
-        shutil.rmtree(output_dir)
-        shutil.rmtree("./artifacts")
+        if os.path.exists(output_dir):
+            shutil.rmtree(output_dir)
+        if os.path.exists("./artifacts"):
+            shutil.rmtree("./artifacts")
